+++ tests/test_configuration.py
@@ -84,12 +84,13 @@ def test_default_configuration_instance(tmp_path):
     ):
         config = Configuration("test.yaml")
         assert config._file_path == tmp_path / "test.yaml"
-        assert "bregenz.a1" in config.profiles
-        assert "eastus2.azure" in config.profiles
-        assert "westeurope.azure" in config.profiles
-        assert config.name == "bregenz.a1"
+        assert "aks1.westeurope.azure" in config.profiles
+        assert "aks1.eastus2.azure" in config.profiles
+        assert "eks1.eu-west-1.aws" in config.profiles
+        assert config.name == "aks1.westeurope.azure"
         assert config.token is None
-        assert config.endpoint == "https://bregenz.a1.cratedb.cloud"
+        assert config.endpoint == "https://console.cratedb.cloud"
+        assert config.region == "aks1.westeurope.azure"
         assert config.format == "table"
         assert config.organization is None
 
@@ -126,8 +127,8 @@ default-format: table
 
 
 def test_use_profile(config):
-    config.use_profile("westeurope.azure")
-    assert config.name == config.config["current-profile"] == "westeurope.azure"
+    config.use_profile("aks1.westeurope.azure")
+    assert config.name == config.config["current-profile"] == "aks1.westeurope.azure"
 
 
 def test_add_profile(config):
@@ -142,12 +143,12 @@ def test_add_profile(config):
 
 def test_add_profile_duplicate(config):
     with pytest.raises(InvalidProfile):
-        config.add_profile("bregenz.a1", endpoint="http://localhost:8000")
+        config.add_profile("aks1.eastus2.azure", endpoint="http://localhost:8000")
 
 
 def test_remove_profile(config):
-    config.remove_profile("eastus2.azure")
-    assert "eastus2.azure" not in config.config["profiles"]
+    config.remove_profile("aks1.eastus2.azure")
+    assert "aks1.eastus2.azure" not in config.config["profiles"]
 
 
 def test_remove_profile_does_not_exist(config):